# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nethesis Operation Center API
  description: REST API for Nethesis Operation Center with business hierarchy management and RBAC
  version: 1.0.0
  contact:
    name: Nethesis S.r.l.
    url: http://www.nethesis.it
    email: info@nethesis.it
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api
    description: Backend API server (port 8080)
  - url: http://localhost:8081/api
    description: Collect API server (port 8081)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from auth exchange endpoint
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication for system inventory collection
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP error code
          example: 400
        message:
          type: string
          description: Error message
          example: "validation failed"
        data:
          $ref: '#/components/schemas/ErrorData'
    ErrorData:
      type: object
      properties:
        type:
          type: string
          enum: [validation_error, external_api_error]
          description: Type of error
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        details:
          description: Additional error details
    ValidationError:
      type: object
      properties:
        key:
          type: string
          description: Field name that failed validation
          example: "username"
        message:
          type: string
          description: Error code or message
          example: "required"
        value:
          type: string
          description: Value that failed validation
          example: ""
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
          example: 20
        total_count:
          type: integer
          minimum: 0
          description: Total number of items
          example: 156
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
          example: 8
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false
        next_page:
          type: integer
          nullable: true
          description: Next page number if available
          example: 2
        prev_page:
          type: integer
          nullable: true
          description: Previous page number if available
          example: null
...
    System:
      type: object
      properties:
        id:
          type: string
          description: System ID
          example: "4cf3053f-d0d5-4b10-b752-ff8f7b63c2f7"
        name:
          type: string
          description: System name
          example: "Production Server 01"
        type:
          type: string
          description: System type
          example: "ns8"
        status:
          type: string
          enum: [online, offline, maintenance]
          description: System operational status
          example: "online"
        fqdn:
          type: string
          description: Fully qualified domain name
          example: "prod-web-01.example.com"
        ipv4_address:
          type: string
          description: IPv4 address
          example: "192.168.1.100"
        ipv6_address:
          type: string
          description: IPv6 address
          example: "2001:db8::1"
        version:
          type: string
          description: System version
          example: "2.1.4"
        last_seen:
          type: string
          format: date-time
          description: Last seen timestamp from inventory
          example: "2025-07-10T10:30:00Z"
        customer_id:
          type: string
          description: Customer organization ID
          example: "org_customer_123"
        custom_data:
          type: object
          description: Custom system data
          additionalProperties: true
          example:
            datacenter: "EU-West-1"
            environment: "production"
            tier: "web"
        secret_hint:
          type: string
          description: Last 4 characters of system secret for identification
          example: "a7b9"
        created_at:
          type: string
          format: date-time
          description: System creation timestamp
          example: "2025-07-01T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: System last update timestamp
          example: "2025-07-10T10:30:00Z"
        created_by:
          $ref: '#/components/schemas/SystemCreator'
        # Heartbeat status fields
        heartbeat_status:
          type: string
          enum: [alive, dead, zombie]
          description: System heartbeat status (alive=<15min, dead=>15min, zombie=never communicated)
          example: "alive"
        last_heartbeat:
          type: string
          format: date-time
          description: Last heartbeat timestamp
          example: "2025-07-21T10:25:00Z"
        heartbeat_minutes:
          type: integer
          description: Minutes since last heartbeat
          example: 5
    SystemCreator:
      type: object
      properties:
        user_id:
          type: string
          description: User ID who created the system
          example: "usr_123456789"
        user_name:
          type: string
          description: Username who created the system
          example: "admin"
        organization_id:
          type: string
          description: Organization ID of the creator
          example: "org_123456789"
        organization_name:
          type: string
          description: Organization name of the creator
          example: "Nethesis"
    SystemInput:
      type: object
      required:
        - name
        - hostname
        - organizationId
      properties:
        name:
          type: string
          description: System name
          example: "Production Server 01"
        description:
          type: string
          description: System description
          example: "Main production server for web applications"
        hostname:
          type: string
          description: System hostname
          example: "prod-web-01.example.com"
        ipAddress:
          type: string
          description: System IP address
          example: "192.168.1.100"
        version:
          type: string
          description: System version
          example: "2.1.4"
        organizationId:
          type: string
          description: Organization ID that owns this system
          example: "org_123456789"
        customData:
          type: object
          description: Custom system data
          additionalProperties: true
          example:
            datacenter: "EU-West-1"
            environment: "production"
            tier: "web"
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchParam:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
        minLength: 1
        example: "acme"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              message:
                type: string
                example: "invalid token"
              data:
                type: object
                nullable: true
                example: null
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 403
              message:
                type: string
                example: "insufficient permissions"
              data:
                type: object
                nullable: true
                example: null
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              message:
                type: string
                example: "internal server error"
              data:
                type: object
                nullable: true
                example: null
paths:
...
  /systems:
    get:
      tags:
        - Backend - Systems
      summary: List systems
      description: Get list of systems (requires Admin or Support role)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: Systems retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "systems retrieved successfully"
                  data:
                    type: object
                    properties:
                      systems:
                        type: array
                        items:
                          $ref: '#/components/schemas/System'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Backend - Systems
      summary: Create system
      description: Create a new system (requires Admin or Support role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemInput'
      responses:
        '201':
          description: System created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "system created successfully"
                  data:
                    $ref: '#/components/schemas/System'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /systems/inventory:
    post:
      servers:
        - url: http://localhost:8081/api
          description: Collect API server (port 8081)
      tags:
        - Collect - Systems
      summary: Collect system inventory
      description: System inventory collection endpoint with HTTP Basic authentication
      security:
        - BasicAuth: [system_id:system_secret]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: System inventory data
                  additionalProperties: true
                  example: {
                    "os": {
                        "name": "my.server",
                        "type": "ns8",
                        "family": "NethServer",
                        "release": {
                            "full": "1.3.5",
                            "major": 1,
                            "minor": 3,
                            "patch": 5
                        }
                    }
                  }
      responses:
        '202':
          description: Inventory received and queued for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Inventory received and queued for processing"
                  data:
                    type: object
                    properties:
                      data_size:
                        type: integer
                        description: Inventory data size in bytes
                        example: 10145
                      message:
                        type: string
                        description: A message from server
                        example: "Your inventory data has been received and will be processed shortly"
                      queue_status:
                        type: string
                        description: Queue status message
                        example: "queued"
                      system_id:
                        type: string
                        description: System ID
                        example: "4cf3053f-d0d5-4b10-b752-ff8f7b63c2f7"
                      timestamp:
                        type: string
                        format: date-time
                        description: Timestamp of received inventory
                        example: "2025-07-16T15:46:51.571831+02:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 413
                  message:
                    type: string
                    example: "Request too large"
                  data:
                    type: object
                    properties:
                      max_size_bytes:
                        type: integer
                        description: Maximum allowed request size in bytes
                        example: 10485760
                      received_bytes:
                        type: integer
                        description: Size of received request in bytes
                        example: 20971520
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Failed to process inventory"
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Processing queue unavailable"
  /systems/heartbeat:
    post:
      servers:
        - url: http://localhost:8081/api
          description: Collect API server (port 8081)
      tags:
        - Collect - Systems
      summary: System heartbeat
      description: System heartbeat endpoint to track system liveness (every 10 minutes)
      security:
        - BasicAuth: [system_id:system_secret]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                system_id:
                  type: string
                  description: System ID sending the heartbeat
                  example: "4cf3053f-d0d5-4b10-b752-ff8f7b63c2f7"
              required:
                - system_id
      responses:
        '200':
          description: Heartbeat acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "heartbeat acknowledged"
                  data:
                    type: object
                    properties:
                      system_id:
                        type: string
                        description: System ID
                        example: "4cf3053f-d0d5-4b10-b752-ff8f7b63c2f7"
                      acknowledged:
                        type: boolean
                        description: Whether heartbeat was acknowledged
                        example: true
                      last_heartbeat:
                        type: string
                        format: date-time
                        description: Timestamp of this heartbeat
                        example: "2025-07-21T10:25:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'